requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
pillow
numpy
faiss-cpu
langchain
sentence-transformers
transformers
